feat:(front)  add NavBar and Footer components, implement TemplateCreation view with image upload and category management, and create NotFound view
