+++ front/src/presentation/components/Footer.vue
@@ -1,0 +1,27 @@
+<template>
+<footer class="border-t-1 border-t-white-custom flex justify-between px-5 h-[200px] py-9">
+  <div class="left flex flex-col text-light-green-custom gap-6">
+    <router-link class="block" to="/">
+      Politique de confidentialité
+    </router-link>
+    <router-link class="block" to="/">
+      Conditions d'utilisation (CGU)
+    </router-link>
+    <p class="text-white-custom">
+      Aucune donnée personnelle collectée • Aucune utilisation de cookies
+    </p>
+  </div>
+  <div class="flex flex-col items-end gap-6 justify-end">
+    <p>
+      contact: lorem ipsum
+    </p>
+    <p>
+      © 2025 TierMaster — Tous droits réservés
+    </p>
+
+  </div>
+
+</footer>
+
+
+</template>
