+++ front/src/presentation/views/NotFoundView.vue
@@ -1,0 +1,3 @@
+<template>
+  404 Not Found
+</template>
